requirements.txt
pdfminer.six
openai
Pandas